requirements/common.txt
#git+https://github.com/kumy/marshmallow-jsonapi.git@missing-relationship#egg=marshmallow-jsonapi&version=0.20.0
git+https://github.com/geokrety/flask-rest-jsonapi.git@geokrety2#egg=flask-rest-jsonapi
git+https://github.com/geokrety/flask.git@issue-2900#egg=Flask&version=1.0.2
marshmallow-jsonapi==0.20.3
gunicorn==19.8.1
Flask-SQLAlchemy==2.3.2
flask-cors==3.0.4
Flask-OAuthlib==0.9.5
flask-login==0.4.1
wtforms==2.2.1
flask-admin==1.5.2
envparse==0.2.0
raven[flask]==6.9.0
flask-jwt==0.3.2
pytz==2018.5
mysqlclient==1.3.13
factory_boy==2.11.1
git+https://github.com/exavolt/python-phpass.git#egg=python-phpass
bcrypt==3.1.4
requests[security]==2.19.1
flask-script==2.0.6
characterentities==0.1.2
geopy==1.16.0
responses==0.9.0
celery[redis]==4.2.1
bleach==2.1.4
